% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/kprod.R
\name{kprod}
\alias{kprod}
\title{Kronecker Product}
\usage{
kprod(...)
}
\arguments{
\item{\ldots}{Either a sequence of matrices or a single list of matrices.}
}
\value{
A matrix with dimensions equal to the product of the input dimensions.
}
\description{
The function \code{kprod} calculates the Kronecker product of two
or more matrices.
}
\examples{
A<-rsan(c(2,3)) ; B<-rsan(c(3,4)) ; C<-rsan(c(4,5))
ABC<-kprod(A,B,C)
dim(ABC)
dim(A)*dim(B)*dim(C)

M<-list(A,B,C)
kprod(M)
}
\author{
Peter Hoff
}

