% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mvreg_ols.R
\name{mvreg_ols}
\alias{mvreg_ols}
\title{OLS Multivariate Linear Regression Estimates}
\usage{
mvreg_ols(Y, X)
}
\arguments{
\item{Y}{An \code{m} by \code{n} matrix.}

\item{X}{A \code{p} by \code{n} matrix.}
}
\value{
A \code{m} by {p} matrix of OLS regression estimates
for the model Y=BX + E.
}
\description{
This functions returns the OLS estimate of the regression
matrix in the multivariate regression model.
}
\examples{
m<-5 ; p<-3 ; n<-10
B<-rsan(c(m,p))
X<-rsan(c(p,n))
E<-rsan(c(m,n))
Y<-B\%*\%X+E

t( lm( t(Y)~ -1 + t(X))$coef )

mvreg_ols(Y,X)
}
\author{
Peter Hoff
}

