% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hmvreg_bayes.R
\name{rB0_fc}
\alias{rB0_fc}
\title{Bayes Hierarchical Multivariate Linear Regression Posterior Simulation}
\usage{
rB0_fc(BA, V0, iV0 = solve(V0))
}
\arguments{
\item{BA}{An \code{r} by \code{m} by \code{p} array.}

\item{V0}{An \code{mp} by \code{mp} matrix.}

\item{iV0}{The inverse of \code{V0}.}
}
\value{
An \code{m} by \code{p} matrix simulated
from the posterior distribution.
}
\description{
This functions returns a simulation from the posterior distribution
of the across-group mean regression coefficient matrix.
}
\author{
Peter Hoff
}

