% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hmvreg_bayes.R
\name{rB_fc}
\alias{rB_fc}
\title{Bayes Multivariate Linear Regression Posterior Simulation}
\usage{
rB_fc(Y, X, B0, V0, VE, YXt = tcrossprod(Y, X), XXt = tcrossprod(X),
  iV0b0 = solve(V0) \%*\% c(B0), iV0 = solve(V0), iVE = solve(VE))
}
\arguments{
\item{Y}{An \code{m} by \code{n} matrix.}

\item{X}{A \code{p} by \code{n} matrix.}

\item{B0}{An \code{m} by \code{p} matrix.}

\item{V0}{A \code{mp} by \code{mp} matrix.}

\item{VE}{An \code{m} by \code{m} matrix.}

\item{YXt}{The value of \code{tcrossprod(Y,X)}.}

\item{XXt}{The value of \code{tcrossprod(X)}.}

\item{iV0b0}{The value of \code{ solve(V0)\%*\%c(B0) }.}

\item{iV0}{The value of \code{ solve(V0) }.}

\item{iVE}{The value of \code{ solve(VE) }.}
}
\value{
An \code{m} by \code{p} matrix simulated
from the posterior distribution.
}
\description{
This functions returns a simulation from the posterior distribution
of the matrix of multivarate linear regression coefficients.
}
\examples{
m<-5 ; p<-3 ; n<-10
B<-rsan(c(m,p))
X<-rsan(c(p,n))
E<-rsan(c(m,n))
Y<-B\%*\%X+E

rB_fc(Y,X,matrix(0,m,p),diag(nrow=m*p),diag(m))

mvreg_ols(Y,X)
}
\author{
Peter Hoff
}

