% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hmvreg_bayes.R
\name{rBA_fc}
\alias{rBA_fc}
\title{Bayes Hierarchical Multivariate Linear Regression Posterior Simulation}
\usage{
rBA_fc(YA, XA, B0, V0, VE)
}
\arguments{
\item{YA}{An \code{r} by \code{m} by \code{n} array.}

\item{XA}{An \code{r} by \code{p} by \code{n} array.}

\item{B0}{An \code{m} by \code{p} matrix.}

\item{V0}{An \code{mp} by \code{mp} matrix.}

\item{VE}{An \code{m} by \code{m} matrix.}
}
\value{
An \code{r} by \code{m} by \code{p} array, simulated
from the posterior distribution.
}
\description{
This functions returns a simulation from the posterior distribution
of the array of matrix of multivarate linear regression coefficients.
}
\author{
Peter Hoff
}

