% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hmvreg_bayes.R
\name{rV0_fc}
\alias{rV0_fc}
\title{Bayes Hierarchical Multivariate Linear Regression Posterior Simulation}
\usage{
rV0_fc(BA, B0)
}
\arguments{
\item{BA}{An \code{r} by \code{m} by \code{p} array.}

\item{B0}{An \code{m} by \code{p} matrix.}
}
\value{
An \code{mp} by \code{mp} matrix simulated
from the posterior distribution.
}
\description{
This functions returns a simulation from the posterior distribution
of the across-group regression coefficient covariance.
}
\author{
Peter Hoff
}

