% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hmvreg_bayes.R
\name{rVE_fc}
\alias{rVE_fc}
\title{Bayes Hierarchical Multivariate Linear Regression Posterior Simulation}
\usage{
rVE_fc(YA, XA, BA)
}
\arguments{
\item{YA}{An \code{r} by \code{m} by \code{n} array.}

\item{XA}{An \code{r} by \code{p} by \code{n} array.}

\item{BA}{An \code{r} by \code{m} by \code{p} array.}
}
\value{
An \code{m} by \code{m} matrix simulated
from the posterior distribution.
}
\description{
This functions returns a simulation from the posterior distribution
of the shared error covariance matrix.
}
\author{
Peter Hoff
}

